+++ src/apps/libraries/ui/library/items/style.scss
@@ -14,4 +14,15 @@
 			}
 		}
 	}
+
+    .fl-asst-import-bar {
+        position: absolute;
+        top: 100%;
+        left: 0;
+        width: 100%;
+        padding: var(--fluid-med-space);
+        background: var(--fluid-opaque-13);
+        border-bottom: 1px solid var(--fluid-opaque-10);
+        color: var(--fluid-color);
+    }
 }
